+++ src/generators/image_generators/stable_diffusion_with_loras.py
@@ -8,8 +8,7 @@ from steamship.data import TagValueKey
 from generators.image_generator import ImageGenerator
 from schema.image_theme import StableDiffusionTheme
 from schema.objects import Item
-from schema.server_settings import get_theme
-from utils.context_utils import get_game_state, get_server_settings
+from utils.context_utils import get_game_state, get_server_settings, get_theme
 from utils.tags import (
     CampTag,
     CharacterTag,
